requirements.txt
crewai
edge_tts==7.0.2
fastapi==0.115.12
whisper==1.1.10
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
streamlit==1.45.1
uvicorn[standard]==0.32.1
python-multipart==0.0.7
aiofiles==24.1.0
pinecone-client>=2.2.4
groq>=0.4.0
yfinance>=0.2.18
beautifulsoup4>=4.12.0
numpy>=1.24.0
torch>=2.1.0
transformers>=4.35.0
